+++ env.example
@@ -5,3 +5,4 @@ IDE_MCP_HOST=http://127.0.0.1:63342/
 MAX_ITERATION=20
 MAX_PROMPT_OUTPUT=8000
 LLM_TEMPERATURE=0.2
+DEBUG=0
